Differentiate examples and tests
